Refactor post component for reusability
