+++ src/app/components/load-posts/load-posts.component.css
@@ -1,0 +1,10 @@
+h2 {
+  display: flex;
+  justify-content: center;
+  font-size: 28px;
+}
+
+#posts {
+  width: 800px;
+  margin-bottom: 30px;
+}
