+++ src/app/components/load-posts/load-posts.component.ts
@@ -7,6 +7,7 @@
 */
 
 import { Component, OnInit } from '@angular/core';
+import { CrudService } from 'src/app/service/crud.service';
 
 @Component({
   selector: 'app-load-posts',
@@ -14,10 +15,15 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./load-posts.component.css']
 })
 export class LoadPostsComponent implements OnInit {
+  OpenPosts:any = []
 
-  constructor() { }
+  constructor(private crudService: CrudService) { }
 
   ngOnInit(): void {
+    this.crudService.GetAllOpenPosts().subscribe(res => {
+      console.log(res)
+      this.OpenPosts = res;
+    })
   }
 
 }
