+++ src/app/components/main/main.component.html
@@ -5,7 +5,7 @@
     <p></p>
 
     <!-- All posts present are for templating only, we will have a loop that displays main and subposts for every parent post. Aren-->
-    <div><app-post></app-post></div>
+    <div><app-load-posts></app-load-posts></div>
   </main>
 
   <aside><app-list5></app-list5></aside>
