+++ src/app/components/main/main.component.ts
@@ -11,6 +11,7 @@ export class MainComponent implements OnInit {
   constructor(private userService: UserService) { }
 
   ngOnInit(): void {
+    
   }
 
 }
