--- src/app/components/main/post/post.component.html
@@ -1,18 +1,0 @@
-<div class="container" *ngFor="let openPost of OpenPosts; let i = index">
-  <div class="header">
-    <div id="post-details">
-      <div id="details">{{ openPost.author }}</div>
-    </div>
-    <div class="menu">
-      <app-kehbab-menu
-        [postID]="openPost._id"
-        [authorID]="openPost.author"
-      ></app-kehbab-menu>
-    </div>
-  </div>
-  <div class="body">
-    <div id="message">{{ openPost.body }}</div>
-    <div id="responses">Sub posts go here</div>
-    <app-reply-post></app-reply-post>
-  </div>
-</div>
