+++ src/app/components/post/post.component.css
@@ -1,7 +1,6 @@
 .container {
-  width: 900px;
+  width: 100%;
   height: 200px;
-  margin: 20px;
   border: 1px solid black;
 }
 
@@ -20,10 +19,17 @@
 
 .body {
   height: 85%;
+  padding: 10px 10px;
 }
 
 #message {
-  margin: 20px;
+  height: 40%;
+  padding-top: 10px;
+  margin-bottom: 10px;
+}
+
+.borderbottom {
+  border-bottom: 1px solid grey;
 }
 
 #post-details {
